move table setup stuff to different file
